refactor: refactor the code for error handling
