+++ src/common/createError.js
@@ -1,5 +1,5 @@
 module.exports = (errorType, message) => {
   const err = new Error(message || 'no text');
   err.status = errorType;
-  throw err;
+  return err;
 };
