+++ src/resources/boards/board.router.js
@@ -40,7 +40,7 @@ router
       const { title, columns } = req.body; // pseudo-validation
 
       if (!title) {
-        createError(
+        throw createError(
           BAD_REQUEST,
           'POST method, enter correct string for board title'
         );
@@ -68,7 +68,10 @@ router
       const boardById = await boardService.getById(id);
 
       if (!boardById) {
-        createError(NOT_FOUND, `GET method, board with ${id} id was not found`);
+        throw createError(
+          NOT_FOUND,
+          `GET method, board with ${id} id was not found`
+        );
       }
 
       res.json(Board.toResponse(boardById));
@@ -81,7 +84,7 @@ router
       const { title, columns } = req.body; // pseudo-validation
 
       if (!title || !columns) {
-        createError(
+        throw createError(
           BAD_REQUEST,
           'PUT method, enter correct string for board title and array for columns'
         );
@@ -102,7 +105,10 @@ router
       });
 
       if (!newBoard) {
-        createError(NOT_FOUND, `PUT method, board with ${id} id was not found`);
+        throw createError(
+          NOT_FOUND,
+          `PUT method, board with ${id} id was not found`
+        );
       }
 
       res.json(Board.toResponse(newBoard));
@@ -116,7 +122,7 @@ router
       const deletedCount = await boardService.deleteBoard(id);
 
       if (deletedCount === 0) {
-        createError(
+        throw createError(
           NOT_FOUND,
           `DELETE method, board with ${id} id was not found`
         );
