+++ src/resources/users/user.router.js
@@ -40,7 +40,7 @@ router
       const { name, login, password } = req.body; // pseudo-validation
 
       if (isEmpty(name) || isEmpty(login) || isEmpty(password)) {
-        createError(
+        throw createError(
           BAD_REQUEST,
           'POST method, enter correct string for user name, login or password'
         );
@@ -61,7 +61,10 @@ router
       const userById = await usersService.getById(id);
 
       if (!userById) {
-        createError(NOT_FOUND, `GET method, user with ${id} id was not found`);
+        throw createError(
+          NOT_FOUND,
+          `GET method, user with ${id} id was not found`
+        );
       }
 
       res.json(User.toResponse(userById));
@@ -75,7 +78,7 @@ router
       const { name, login, password } = req.body; // pseudo-validation
 
       if (isEmpty(name) || isEmpty(login) || isEmpty(password)) {
-        createError(
+        throw createError(
           BAD_REQUEST,
           'PUT method, enter correct string for user name, login or password'
         );
@@ -84,7 +87,10 @@ router
       const user = await usersService.update({ id, name, login, password });
 
       if (!user) {
-        createError(NOT_FOUND, `PUT method, user with ${id} id was not found`);
+        throw createError(
+          NOT_FOUND,
+          `PUT method, user with ${id} id was not found`
+        );
       }
 
       res.json(User.toResponse(user));
@@ -98,7 +104,7 @@ router
       const deletedCount = await usersService.deleteUser(id);
 
       if (deletedCount === 0) {
-        createError(
+        throw createError(
           NOT_FOUND,
           `DELETE method, user with ${id} id was not found`
         );
